+++ catsup/cli.py
@@ -32,6 +32,7 @@ Usage:
     catsup server [-s <file>|--settings=<file>] [-p <port>|--port=<port>]
     catsup webhook [-s <file>|--settings=<file>] [-p <port>|--port=<port>]
     catsup themes
+    catsup clean [-s <file>|--settings=<file>]
     catsup install <theme>
     catsup -h | --help
     catsup --version
@@ -174,6 +175,28 @@ def themes():
     """
     import catsup.parser.themes
     catsup.parser.themes.list()
+
+
+@parguments.command
+def clean(settings):
+    """
+    Usage:
+        catsup clean [-s <file>|--settings=<file>]
+
+    Options:
+        -s --settings=<file>    specify a setting file. [default: config.json]
+        -h --help               Show this screen and exit.
+    """
+    import catsup.parser.config
+    import shutil
+    config = catsup.parser.config(settings)
+    if os.path.exists(".catsup-cache"):
+        shutil.rmtree(".catsup-cache")
+        logger.info("Removed cache folder.")
+    output_path = config.config.output
+    if os.path.exists(output_path):
+        shutil.rmtree(output_path)
+        logger.info("Removed output folder.")
 
 
 @parguments.command